renamed SCons.Tool.ninja -> SCons.Tool.ninja_tool and added alias in tool loading logic.  This fixes changes in this PR breaking JavaCommonTests because pypi's ninja module and SCons.Tool.ninja had the same name which python couldn't differentiate
