+++ SCons/Tool/ninja_tool/Globals.py
