+++ SCons/Tool/ninja_tool/Overrides.py
