+++ SCons/Tool/ninja_tool/Rules.py
