+++ SCons/Tool/ninja_tool/ninja.xml
