+++ SCons/Tool/ninja_tool/ninja_daemon_build.py
