+++ SCons/Tool/ninja_tool/ninja_run_daemon.py
