+++ SCons/Tool/ninja_tool/ninja_scons_daemon.py
